test(ui): Update tests according to `RoomDataProvider` recent change.
