+++ crates/matrix-sdk-ui/src/timeline/event_item/mod.rs
@@ -468,12 +468,10 @@ pub enum EventItemOrigin {
 #[cfg(test)]
 mod tests {
     use assert_matches::assert_matches;
-    use matrix_sdk::{config::RequestConfig, Client, ClientBuilder};
-    use matrix_sdk_base::{deserialized_responses::SyncTimelineEvent, BaseClient, SessionMeta};
+    use matrix_sdk_base::deserialized_responses::SyncTimelineEvent;
     use matrix_sdk_test::async_test;
     use ruma::{
-        api::{client::sync::sync_events::v4, MatrixVersion},
-        device_id,
+        api::client::sync::sync_events::v4,
         events::{
             room::message::{MessageFormat, MessageType},
             AnySyncTimelineEvent,
@@ -485,7 +483,7 @@ mod tests {
     use serde_json::json;
 
     use super::{EventTimelineItem, Profile};
-    use crate::timeline::TimelineDetails;
+    use crate::timeline::{tests::logged_in_client, TimelineDetails};
 
     #[async_test]
     async fn latest_message_event_can_be_wrapped_as_a_timeline_item() {
@@ -609,29 +607,5 @@ mod tests {
             )
             .unwrap(),
         )
-    }
-
-    /// Copied from matrix_sdk_base::sliding_sync::test
-    async fn logged_in_client(homeserver_url: Option<String>) -> Client {
-        let base_client = BaseClient::new();
-        base_client
-            .set_session_meta(SessionMeta {
-                user_id: user_id!("@u:e.uk").to_owned(),
-                device_id: device_id!("XYZ").to_owned(),
-            })
-            .await
-            .expect("Failed to set session meta");
-
-        test_client_builder(homeserver_url)
-            .request_config(RequestConfig::new().disable_retry())
-            .base_client(base_client)
-            .build()
-            .await
-            .unwrap()
-    }
-
-    fn test_client_builder(homeserver_url: Option<String>) -> ClientBuilder {
-        let homeserver = homeserver_url.as_deref().unwrap_or("http://localhost:1234");
-        Client::builder().homeserver_url(homeserver).server_versions([MatrixVersion::V1_0])
     }
 }
