+++ crates/matrix-sdk-ui/src/timeline/tests/reactions.rs
@@ -40,7 +40,7 @@ const REACTION_KEY: &str = "👍";
 
 #[async_test]
 async fn add_reaction_failed() {
-    let timeline = TestTimeline::new();
+    let timeline = TestTimeline::new().await;
     let mut stream = timeline.subscribe().await;
     let (msg_id, msg_pos) = send_first_message(&timeline, &mut stream).await;
     let reaction = create_reaction(&msg_id);
@@ -60,7 +60,7 @@ async fn add_reaction_failed() {
 
 #[async_test]
 async fn add_reaction_on_non_existent_event() {
-    let timeline = TestTimeline::new();
+    let timeline = TestTimeline::new().await;
     let mut stream = timeline.subscribe().await;
     let msg_id = EventId::new(server_name!("example.org")); // non existent event
     let reaction = create_reaction(&msg_id);
@@ -72,7 +72,7 @@ async fn add_reaction_on_non_existent_event() {
 
 #[async_test]
 async fn add_reaction_success() {
-    let timeline = TestTimeline::new();
+    let timeline = TestTimeline::new().await;
     let mut stream = timeline.subscribe().await;
     let (msg_id, msg_pos) = send_first_message(&timeline, &mut stream).await;
     let reaction = create_reaction(&msg_id);
@@ -96,7 +96,7 @@ async fn add_reaction_success() {
 
 #[async_test]
 async fn redact_reaction_success() {
-    let timeline = TestTimeline::new();
+    let timeline = TestTimeline::new().await;
     let mut stream = timeline.subscribe().await;
     let (msg_id, msg_pos) = send_first_message(&timeline, &mut stream).await;
     let reaction = create_reaction(&msg_id);
@@ -118,7 +118,7 @@ async fn redact_reaction_success() {
 
 #[async_test]
 async fn redact_reaction_failure() {
-    let timeline = TestTimeline::new();
+    let timeline = TestTimeline::new().await;
     let mut stream = timeline.subscribe().await;
     let (msg_id, msg_pos) = send_first_message(&timeline, &mut stream).await;
     let reaction = create_reaction(&msg_id);
@@ -144,7 +144,7 @@ async fn redact_reaction_failure() {
 
 #[async_test]
 async fn redact_reaction_from_non_existent_event() {
-    let timeline = TestTimeline::new();
+    let timeline = TestTimeline::new().await;
     let mut stream = timeline.subscribe().await;
     let reaction_id = EventId::new(server_name!("example.org")); // non existent event
 
@@ -157,7 +157,7 @@ async fn redact_reaction_from_non_existent_event() {
 
 #[async_test]
 async fn toggle_during_request_resolves_new_action() {
-    let timeline = TestTimeline::new();
+    let timeline = TestTimeline::new().await;
     let mut stream = timeline.subscribe().await;
     let (msg_id, msg_pos) = send_first_message(&timeline, &mut stream).await;
     let reaction = create_reaction(&msg_id);
@@ -208,7 +208,7 @@ async fn toggle_during_request_resolves_new_action() {
 
 #[async_test]
 async fn reactions_store_timestamp() {
-    let timeline = TestTimeline::new();
+    let timeline = TestTimeline::new().await;
     let mut stream = timeline.subscribe().await;
     let (msg_id, msg_pos) = send_first_message(&timeline, &mut stream).await;
     let reaction = create_reaction(&msg_id);
@@ -241,7 +241,7 @@ async fn reactions_store_timestamp() {
 
 #[async_test]
 async fn initial_reaction_timestamp_is_stored() {
-    let mut timeline = TestTimeline::new();
+    let mut timeline = TestTimeline::new().await;
 
     let message_event_id = EventId::new(server_name!("dummy.server"));
     let reaction_timestamp = MilliSecondsSinceUnixEpoch(uint!(39845));
